Feat - SRP on PostController
더 이상 PostController에서 PostRepository를injection 받지 않습니다.

기타 로직들이 PostService로 이전되었습니다.

+++ src/main/java/onboarding/crud/service/PostService.java
@@ -31,7 +31,29 @@ public class PostService {
         return postRepository.findAll();
     }
 
-    public void deletePost(Long id) throws Exception {
-        postRepository.deleteById(id);
+    public boolean deletePost(Long id) {
+        if(postRepository.existsById(id)){
+            postRepository.deleteById(id);
+            return true;
+        }else {
+            return false;
+        }
+    }
+
+    public boolean modifyPost(Long id, Post newPost) {
+        Optional<Post> oldPost = postRepository.findById(id);
+        if (oldPost.isPresent()) {
+            Post updatedPost = oldPost.get();
+            if (newPost.getTitle() != null) {
+                updatedPost.setTitle(newPost.getTitle());
+            }
+            if (newPost.getContent() != null) {
+                updatedPost.setContent(newPost.getContent());
+            }
+            postRepository.save(updatedPost);
+            return true;
+        } else {
+            return false;
+        }
     }
 }
